Plant Leaf Disease Detection Using CNN/requirements.txt
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
streamlit==1.22.0
tensorflow-cpu==2.12.0

# Commented out older versions for reference
# keras==2.8.0rc0
# numpy==1.20.1
# Pillow==8.4.0
# streamlit==1.2.0
# tensorflow-cpu==2.8.0rc0